[monorepo]: move voting key generation tests to test_vectors
 problem: there are multiple copies of voting key generation test vectors
          and they are being used during unit tests
solution: promote vector file to tests/vectors
          process it as part of test_vectors stage
