+++ sdk/javascript/package.json
@@ -24,7 +24,6 @@
     "mocha": "^9.2.2",
     "mocha-jenkins-reporter": "^0.4.7",
     "nyc": "^15.1.0",
-    "yaml": "^2.1.1",
     "yargs": "^17.5.1"
   },
   "dependencies": {
