+++ sdk/python/tests/vectors/all.py
@@ -9,6 +9,7 @@ from symbolchain.Bip32 import Bip32
 from symbolchain.CryptoTypes import PrivateKey, PublicKey, Signature
 from symbolchain.Network import NetworkLocator
 from symbolchain.symbol.IdGenerator import generate_mosaic_id
+from symbolchain.symbol.VotingKeysGenerator import VotingKeysGenerator
 
 
 class ClassLocator:
@@ -24,7 +25,7 @@ class ClassLocator:
 		return Bip32(self.facade_class.BIP32_CURVE_NAME)
 
 
-# region TestSuites
+# region VectorsTestSuite, KeyConversionTester, AddressConversionTester
 
 class VectorsTestSuite:
 	def __init__(self, identifier, filename, description):
@@ -78,6 +79,10 @@ class AddressConversionTester(VectorsTestSuite):
 			(expected_address_testnet, actual_address_testnet)
 		]
 
+# endregion
+
+
+# region SignTester, VerifyTester
 
 class SignTester(VectorsTestSuite):
 	def __init__(self, class_locator):
@@ -114,6 +119,10 @@ class VerifyTester(VectorsTestSuite):
 		# Assert:
 		return [(True, is_verified)]
 
+# endregion
+
+
+# region MosaicIdDerivationTester
 
 class MosaicIdDerivationTester(VectorsTestSuite):
 	def __init__(self, class_locator):
@@ -137,6 +146,10 @@ class MosaicIdDerivationTester(VectorsTestSuite):
 
 		return mosaic_id_pairs
 
+# endregion
+
+
+# region Bip32DerivationTester, Bip39DerivationTester
 
 class Bip32DerivationTester(VectorsTestSuite):
 	def __init__(self, class_locator):
@@ -190,8 +203,68 @@ class Bip39DerivationTester(VectorsTestSuite):
 		# Assert:
 		return [(expected_root_public_key, root_public_key)]
 
+# endregion
+
+
+# region VotingKeysGenerationTester
+
+class SeededPrivateKeyGenerator:
+	def __init__(self, values):
+		self.values = values
+		self.next_index = 0
+
+	def generate(self):
+		self.next_index += 1
+		return self.values[self.next_index - 1]
+
+
+class FibPrivateKeyGenerator:
+	def __init__(self, fill_private_key=False):
+		self.fill_private_key = fill_private_key
+		self.value1 = 1
+		self.value2 = 2
+
+	def generate(self):
+		next_value = self.value1 + self.value2
+		self.value1 = self.value2
+		self.value2 = next_value
+
+		seed_value = next_value % 256
+
+		if not self.fill_private_key:
+			return PrivateKey(seed_value.to_bytes(PrivateKey.SIZE, 'big'))
+
+		return PrivateKey(bytes([(seed_value + i) % 256 for i in range(0, PrivateKey.SIZE)]))
+
+
+class VotingKeysGenerationTester(VectorsTestSuite):
+	def __init__(self, class_locator):
+		super().__init__(7, 'test-voting-keys-generation', 'voting keys generation')
+		self.class_locator = class_locator
+
+	def process(self, test_vector, _):
+		private_key_generator = {
+			'test_vector_1': FibPrivateKeyGenerator(),
+			'test_vector_2': FibPrivateKeyGenerator(True),
+			'test_vector_3': SeededPrivateKeyGenerator([
+				PrivateKey('12F98B7CB64A6D840931A2B624FB1EACAFA2C25C3EF0018CD67E8D470A248B2F'),
+				PrivateKey('B5593870940F28DAEE262B26367B69143AD85E43048D23E624F4ED8008C0427F'),
+				PrivateKey('6CFC879ABCCA78F5A4C9739852C7C643AEC3990E93BF4C6F685EB58224B16A59')
+			])
+		}[test_vector['name']]
+
+		root_private_key = PrivateKey(test_vector['rootPrivateKey'])
+		voting_keys_generator = VotingKeysGenerator(self.class_locator.key_pair_class(root_private_key), private_key_generator.generate)
+
+		# Act:
+		voting_keys_buffer = voting_keys_generator.generate(test_vector['startEpoch'], test_vector['endEpoch'])
+
+		# Assert:
+		expected_voting_keys_buffer = unhexlify(test_vector['expectedFileHex'])
+		return [(expected_voting_keys_buffer, voting_keys_buffer)]
 
 # endregion
+
 
 def load_class_locator(blockchain):
 	# pylint: disable=import-outside-toplevel
@@ -218,7 +291,7 @@ def load_test_suites(blockchain):
 	]
 
 	if 'symbol' == blockchain:
-		test_suites += [MosaicIdDerivationTester(class_locator)]
+		test_suites += [MosaicIdDerivationTester(class_locator), VotingKeysGenerationTester(class_locator)]
 
 	return test_suites
 
@@ -226,7 +299,7 @@ def load_test_suites(blockchain):
 def main():
 	# pylint: disable=too-many-locals
 
-	test_identifiers = range(0, 7)
+	test_identifiers = range(0, 8)
 	parser = argparse.ArgumentParser(description='nem test vectors harness')
 	parser.add_argument('--vectors', help='path to test-vectors directory', required=True)
 	parser.add_argument('--blockchain', help='blockchain to run vectors against', choices=['nem', 'symbol'], default='symbol')
